Make `Cardano.Wallet.Read.Block` a hierarchy root
