+++ lib/read/lib/Cardano/Wallet/Read/Block/Txs.hs
@@ -11,7 +11,7 @@ import Prelude
 import Cardano.Ledger.Binary
     ( EncCBOR
     )
-import Cardano.Wallet.Read.Block
+import Cardano.Wallet.Read.Block.Block
     ( Block (..)
     )
 import Cardano.Wallet.Read.Eras
